linux-renesas: fix wrong config problem
A race condition may happen. When it happens, the configs
provided in defconfig file are not merged to .config
Thus the built image will not have neccessary configs
and cannot boot up.

This commit is created based on poky commit 81e8a5
"kernel: specify dependencies for compilation for config tasks"

Signed-off-by: Hung Tran <hung.tran.jy@renesas.com>