SDK: fix error if host system does not support --sysroot
Certain system like Asianux Server 7 does not support --sysroot option.
Thus error will happen when install current SDK on that system.
This commit modifies the environment-setup scrip to avoid this kind of error.
Note that with this change, when install SDK on Ubuntu 14.04, many warning
message will appear. It is safe to ignore these warning messages.

Signed-off-by: Hung Tran <hung.tran.jy@renesas.com>