weston: weston_v4l2.ini: disable LVDS-2 output
Due to dual channels LVDS, it will create 2 panel LVDS-1 and LVDS-2.
LVDS-1 is in our main panel, LVDS-2 is on the right of panel.
But we just use LVDS-1 to display for both channels.

So that turn off LVDS-2 to avoid images/videos or applications can
display in LVDS-2.

Signed-off-by: Hien Huynh <hien.huynh.px@renesas.com>