bluez5,ofono: remove RDEPEND of python3-pygobject
python3-pygobject depends on cairo, then cairo depends on
gles-user-module which is not included in core-image-minimal|bsp.
So minimal and bsp image will be failed to build.

This commit remove RDEPEND of python3-pygobject in bluez5, ofono to
prevent the build failure. python3-pygobject is just depended in case
bluez5 and ofono enable test/test-tool

Signed-off-by: Binh Nguyen <binh.nguyen.uw@renesas.com>