cip-core: Fix and improve settings when not use cip-core
For building without packages in recipes-cip-core, user needs to specify
the PREFERRED_VERSION of default packages in poky.
However, it is troublesome when poky change the packages version.

This improvement can solve that situation by MASK out recipes-cip-core
when it is not used. Then Yocto will ignores our packages in recipe-cip-core.

Signed-off-by: Hung Tran <hung.tran.jy@renesas.com>