cip-core-buster: Fix build for libxml-parser-perl without new classes
Previously libxml-parser-perl recipe (Buster version) requires user to update
some related classes in poky to be able to build correctly.
It is inconvenient, and also cause issue if user changes build setting.

After this fix, libxml-parser-perl Buster version can be built with current
classes in poky. To do that, a few overwrite settings are added to recipe
to correct the problems due to mismatch of the old classes.

Signed-off-by: Hung Tran <hung.tran.jy@renesas.com>