recipes-multimedia: gstreamer: plugins-good: v4l2src: Ensure all caps a fixated
The code relied on the list compare function to fixate the caps
but if the caps only has one structure, the compare function will
never get called. Capture device for which there is only one
structure in the caps would then get some assertion and later
fail badly.

Instead, fixate before inserting into the list and split the reading
and the fixation of the structures.

Reference:
https://github.com/GStreamer/gst-plugins-good/commit/a08d7cdef5e20b88f93674a5b1053dc6edee7e18

Signed-off-by: Huynh Thanh Hung <hung.huynh.wz@renesas.com>