docs/sample/conf: Avoid error in Gecko Jessie with version of nodejs and libgit2
Recent update causes error when build Gecko with Jessie.
This happens because of new versions of nodejs and libgit2 are not compatible
with old openssl version.

Fix by adjust the nodejs and libgit2 according to CIP_MODE.
Also apply to non Gecko build, in case user wants to use these packages.

Signed-off-by: Hung Tran <hung.tran.jy@renesas.com>