recipes-bsp: arm-trusted-firmware: add init counter wdt and re-work initialize memory when enable ECC
This commit will add the init counter function for watchdog, so that
other process can reset the watchdog counter.

Previous implement can cause the missing initialize memory of FUSA/ECC
area when changing the configuration, and make the ECC error occur.
This time we will initialize the memory base on the configuration to
avoid missing.

The new implement also use reset counter of watchdog for each erase
memory area, so that no need to increase the watchdog counter anymore.

Signed-off-by: Huynh Thanh Hung <hung.huynh.wz@renesas.com>